first script
I wrote this today as the start of this project. It's a single script implementation of Splunk custom search command (version 2). In an earlier experiment I figured out how to contain an entire Python3 directory into a Splunk App and reroute the stdin/stdout to that alternate Python. But... it was bulky and hacky. I'd prefer if the code were backwards compatible with Python2, but written in Python3. Also, if the code could use built-in libraries from Python then that would really trim down the size of the app, but I know that may not be possible for everything. (Python3's import engine is so jank though...)