built in python3 with much better features/speed
This is now pretty much an SDK in and of itself. It has the ability to be any kind of command dynamically (decided during search time). It parses the command arguments into a globally available dictionary (see line 98) for structure. It actively converts MV fields into lists and back to Splunk appropriate MV syntax on the way back, and it can change field names/order mid-processing. But it also has the ability to turn those extra processes off to make it much faster. Also for speed, it pipes the body from Splunk directly into a csv writer as it processes each row (unless field name alterations are enabled). I'd love to put it straight into stdin, but you have to preface your chunk with the body length, which is only known after you get done processing your chunk. It also has (albeit dry and untested) methods for generating new events, and it will decide on either generate or processing rows based on the settings.
I still need to annotate it, but I also wanted to say that I tried recreating this script in C# and I couldn't get it to run significantly faster than python. Because Python is so intimately connected to Splunk, I decided beforehand that I would only change the language if I could get about a 50% drop in runtime. I may try some other languages down the line, but I have pretty high regard for C#'s performance, so I'm not holding out much hope. To be fair, though, I was only testing based on the stdin/stdout cycle between this script's process and Splunk's engine. Other languages might realize much more appealing performance boosts when it comes to the actual row-by-row data processing. (Especially when you account for Python's limitations with threading) So I wouldn't say it's completely over for other languages...